gsplat-rust-renderer: dedupe camera loaders + reuse simplecv up-vector helper
Simplify pass over the relocated apis/ CLIs (parallel reuse/quality/efficiency review):
- Extract the duplicated COLMAP binary readers, qvec->rotmat, NeRF-synthetic loader, and
  RGBA->RGB compositing into gsplat_rust_renderer/scene_io.py; both visualize_brush_training
  and log_splats_with_cameras import them. The copies had already drifted (only one
  supported OPENCV_FISHEYE) — scene_io uses the superset model table.
- Replace the hand-rolled up_to_z_rotation Rodrigues with simplecv's existing
  rotation_matrix_between (numerically identical; simplecv is already a dep).
- count_eval_views: bind colmap_sparse_dir once instead of calling it twice + asserting.

load_colmap_cameras kept per-file (the two genuinely differ: nested-dir + downscaled
thumbnails + distortion-tolerant slicing vs sparse/0 + full-res + 4-param), so behavior is
unchanged; both now call the shared readers. Kept the hand-rolled qvec_to_rotmat (COLMAP
wxyz vs scipy xyzw is a footgun, and it is not hot-path).

visualize_brush_training 835->726, log_splats 284->178 lines. Verified: ruff + pyrefly
clean, 13 tests pass, all 3 CLIs --help, apis + scene_io import under beartype.
