+++ packages/gsplat-rust-renderer/gsplat_rust_renderer/apis/log_splats_with_cameras.py
@@ -26,8 +26,6 @@ Examples:
         --eye 2.84 -4.09 -6.12 --look-target 0.43 -0.07 0.31 --eye-up 0 -1 0
 """
 
-import json
-import struct
 from dataclasses import dataclass
 from pathlib import Path
 from typing import Literal
@@ -35,13 +33,14 @@ from typing import Literal
 import numpy as np
 import rerun as rr
 import rerun.blueprint as rrb
-from jaxtyping import Float64, UInt8
+from jaxtyping import UInt8
 from numpy import ndarray
 from PIL import Image
 from simplecv.camera_parameters import Extrinsics, Intrinsics, PinholeParameters
 from simplecv.rerun_log_utils import RerunTyroConfig, log_pinhole
 
 from gsplat_rust_renderer.gaussians3d import Gaussians3D
+from gsplat_rust_renderer.scene_io import load_nerf_cameras, load_rgb_composited, qvec_to_rotmat, read_colmap_cameras_bin, read_colmap_images_bin
 
 
 @dataclass
@@ -74,91 +73,6 @@ class LogSplatsWithCamerasConfig:
     """Up vector for the initial eye pose (used only when --eye is given)."""
 
 
-def load_nerf_cameras(scene_dir: Path) -> list[tuple[PinholeParameters, Path]]:
-    """Read NeRF-synthetic test cameras as (pinhole, image path) pairs.
-
-    The c2w matrices are OpenGL/RUB; simplecv carries the convention through
-    to the Pinhole's ``camera_xyz``, so they are used unmodified.
-    """
-    transforms = json.loads((scene_dir / "transforms_test.json").read_text())
-    camera_angle_x: float = transforms["camera_angle_x"]
-    cameras: list[tuple[PinholeParameters, Path]] = []
-    for frame in transforms["frames"]:
-        image_path: Path = (scene_dir / frame["file_path"]).with_suffix(".png")
-        with Image.open(image_path) as probe:
-            width, height = probe.size
-        focal: float = 0.5 * width / np.tan(0.5 * camera_angle_x)
-        c2w: Float64[ndarray, "4 4"] = np.array(frame["transform_matrix"], dtype=np.float64)
-        camera = PinholeParameters(
-            name=image_path.stem,
-            extrinsics=Extrinsics(world_R_cam=c2w[:3, :3], world_t_cam=c2w[:3, 3]),
-            intrinsics=Intrinsics.from_focal_principal_point(
-                camera_conventions="RUB",
-                fl_x=focal,
-                fl_y=focal,
-                cx=width / 2.0,
-                cy=height / 2.0,
-                height=height,
-                width=width,
-            ),
-        )
-        cameras.append((camera, image_path))
-    return cameras
-
-
-# ── Minimal COLMAP binary parsers ────────────────────────────────────────────
-# pycolmap is not in this package's environment; the binary layout is stable
-# and these ~40 lines avoid a heavyweight dependency for a demo tool.
-
-
-def _read_cameras_bin(path: Path) -> dict[int, dict]:
-    models: dict[int, tuple[str, int]] = {
-        0: ("SIMPLE_PINHOLE", 3),
-        1: ("PINHOLE", 4),
-        2: ("SIMPLE_RADIAL", 4),
-        3: ("RADIAL", 5),
-        4: ("OPENCV", 8),
-    }
-    cameras: dict[int, dict] = {}
-    with open(path, "rb") as f:
-        (n,) = struct.unpack("<Q", f.read(8))
-        for _ in range(n):
-            cam_id, model_id, w, h = struct.unpack("<iiQQ", f.read(24))
-            _name, n_params = models[model_id]
-            params = struct.unpack(f"<{n_params}d", f.read(8 * n_params))
-            cameras[cam_id] = {"width": w, "height": h, "params": params}
-    return cameras
-
-
-def _read_images_bin(path: Path) -> list[dict]:
-    images: list[dict] = []
-    with open(path, "rb") as f:
-        (n,) = struct.unpack("<Q", f.read(8))
-        for _ in range(n):
-            f.read(4)  # image id
-            qvec = struct.unpack("<4d", f.read(32))
-            tvec = struct.unpack("<3d", f.read(24))
-            cam_id = struct.unpack("<i", f.read(4))[0]
-            name = b""
-            while (c := f.read(1)) != b"\x00":
-                name += c
-            (n_pts,) = struct.unpack("<Q", f.read(8))
-            f.read(24 * n_pts)  # skip 2D points
-            images.append({"qvec": qvec, "tvec": tvec, "camera_id": cam_id, "name": name.decode()})
-    return images
-
-
-def _qvec_to_rotmat(q: tuple[float, float, float, float]) -> Float64[ndarray, "3 3"]:
-    w, x, y, z = q
-    return np.array(
-        [
-            [1 - 2 * (y * y + z * z), 2 * (x * y - w * z), 2 * (x * z + w * y)],
-            [2 * (x * y + w * z), 1 - 2 * (x * x + z * z), 2 * (y * z - w * x)],
-            [2 * (x * z - w * y), 2 * (y * z + w * x), 1 - 2 * (x * x + y * y)],
-        ]
-    )
-
-
 def load_colmap_cameras(scene_dir: Path) -> list[tuple[PinholeParameters, Path]]:
     """Read COLMAP sparse cameras as (pinhole, image path) pairs.
 
@@ -167,8 +81,8 @@ def load_colmap_cameras(scene_dir: Path) -> list[tuple[PinholeParameters, Path]]
     Intrinsics are rescaled when the shipped images are smaller than the
     calibrated resolution (T&T ships half-res images).
     """
-    calibrations = _read_cameras_bin(scene_dir / "sparse" / "0" / "cameras.bin")
-    poses = sorted(_read_images_bin(scene_dir / "sparse" / "0" / "images.bin"), key=lambda im: im["name"])
+    calibrations = read_colmap_cameras_bin(scene_dir / "sparse" / "0" / "cameras.bin")
+    poses = sorted(read_colmap_images_bin(scene_dir / "sparse" / "0" / "images.bin"), key=lambda im: im["name"])
     cameras: list[tuple[PinholeParameters, Path]] = []
     for im in poses:
         calib = calibrations[im["camera_id"]]
@@ -181,7 +95,7 @@ def load_colmap_cameras(scene_dir: Path) -> list[tuple[PinholeParameters, Path]]
         camera = PinholeParameters(
             name=Path(im["name"]).stem,
             extrinsics=Extrinsics(
-                cam_R_world=_qvec_to_rotmat(im["qvec"]),
+                cam_R_world=qvec_to_rotmat(im["qvec"]),
                 cam_t_world=np.asarray(im["tvec"], dtype=np.float64),
             ),
             intrinsics=Intrinsics.from_focal_principal_point(
@@ -198,26 +112,9 @@ def load_colmap_cameras(scene_dir: Path) -> list[tuple[PinholeParameters, Path]]
     return cameras
 
 
-def load_rgb_on_white(image_path: Path) -> UInt8[ndarray, "h w 3"]:
-    """Load a GT image as raw RGB, compositing any alpha onto white (the
-    NeRF-synthetic images are RGBA over transparent; the splat renders use a
-    white background).
-
-    Raw ``rr.Image`` on purpose: ``rr.EncodedImage`` makes the viewer
-    re-decode every visible JPEG every frame (decode cache misses) — 200 image
-    planes pegged a core and dropped frames to seconds.  Raw rows cost more
-    memory but render at 60 FPS.
-    """
-    with Image.open(image_path) as img:
-        rgba: UInt8[ndarray, "h w c"] = np.asarray(img.convert("RGBA"))
-    alpha: Float64[ndarray, "h w 1"] = rgba[..., 3:].astype(np.float64) / 255.0
-    rgb: UInt8[ndarray, "h w 3"] = (rgba[..., :3].astype(np.float64) * alpha + 255.0 * (1.0 - alpha)).astype(np.uint8)
-    return rgb
-
-
 def main(config: LogSplatsWithCamerasConfig) -> None:
     if (config.scene_dir / "transforms_test.json").exists():
-        cameras: list[tuple[PinholeParameters, Path]] = load_nerf_cameras(config.scene_dir)
+        cameras: list[tuple[PinholeParameters, Path]] = load_nerf_cameras(config.scene_dir, "test")
     elif (config.scene_dir / "sparse" / "0" / "cameras.bin").exists():
         cameras = load_colmap_cameras(config.scene_dir)
     else:
@@ -231,7 +128,7 @@ def main(config: LogSplatsWithCamerasConfig) -> None:
     for k, (camera, image_path) in enumerate(cameras):
         cam_path = f"world/cameras/{camera.name}"
         log_pinhole(camera, cam_log_path=Path(cam_path), image_plane_distance=config.image_plane_distance, static=True)
-        rgb: UInt8[ndarray, "h w 3"] = load_rgb_on_white(image_path)
+        rgb: UInt8[ndarray, "h w 3"] = load_rgb_composited(image_path, 255.0)
         rr.log(f"{cam_path}/pinhole/image", rr.Image(rgb), static=True)
         rr.set_time("page", sequence=k // 4)
         rr.log(f"browser/{k % 4}", rr.Image(rgb))
